Wrap header fields with an indexer for performance

+++ MatchRule.cs
@@ -148,7 +148,7 @@ namespace NDesk.DBus
 
 			foreach (KeyValuePair<FieldCode,MatchTest> pair in Fields) {
 				object value;
-				if (!msg.Header.Fields.TryGetValue (pair.Key, out value))
+				if (!msg.Header.Fields.TryGetValue ((byte)pair.Key, out value))
 					return false;
 				if (!pair.Value.Value.Equals (value))
 					return false;
